move formatting utils to util package
