+++ src/main/java/org/rumbledb/runtime/functions/util/formatting/pictures/FormatNumber/FormatNumberPictureSupport.java
@@ -1,7 +1,7 @@
-package org.rumbledb.runtime.functions.base.formatting.pictures.FormatNumber;
+package org.rumbledb.runtime.functions.util.formatting.pictures.FormatNumber;
 
 import org.rumbledb.context.DecimalFormatDefinition;
-import org.rumbledb.runtime.functions.base.formatting.GroupingPos;
+import org.rumbledb.runtime.functions.util.formatting.GroupingPos;
 
 import java.util.ArrayList;
 import java.util.Comparator;
