+++ src/main/java/org/rumbledb/runtime/functions/util/formatting/pictures/FormatNumber/FormatNumberSubPicture.java
@@ -1,6 +1,6 @@
-package org.rumbledb.runtime.functions.base.formatting.pictures.FormatNumber;
+package org.rumbledb.runtime.functions.util.formatting.pictures.FormatNumber;
 
-import org.rumbledb.runtime.functions.base.formatting.GroupingPos;
+import org.rumbledb.runtime.functions.util.formatting.GroupingPos;
 
 import java.util.List;
