+++ src/main/java/org/rumbledb/runtime/functions/util/formatting/timezone/BuiltinTimezoneNameFormatter.java
@@ -1,4 +1,4 @@
-package org.rumbledb.runtime.functions.base.formatting.timezone;
+package org.rumbledb.runtime.functions.util.formatting.timezone;
 
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
