+++ src/main/java/org/rumbledb/runtime/functions/util/formatting/timezone/TimezoneNameContext.java
@@ -1,4 +1,4 @@
-package org.rumbledb.runtime.functions.base.formatting.timezone;
+package org.rumbledb.runtime.functions.util.formatting.timezone;
 
 import java.time.ZoneId;
 import java.util.Locale;
